More appropriate naming. Check for host.id existence and bail if needed.

+++ scripts/run.sh
@@ -1,16 +1,18 @@
 #!/bin/bash
 
 DIR=`pwd`
-CURRENT_DIR="${DIR##*/}"
+CONTAINER="${DIR##*/}"
 
 # The docker index server can be set to a local index, or by default it will
 # use the current directory name before the `-` separator.
 #    for example: datt-nginx will use the `datt` index account.
 
-INDEX_NAME=$(echo $CURRENT_DIR | cut -d- -f 1)
+INDEX_NAME=$(echo $CONTAINER | cut -d- -f 1)
 DOCKERINDEX=${DOCKERINDEX_LOCAL:-"$INDEX_NAME/"}
 
 DENV=`test $# -gt 0 && echo "--env=$*" || echo ''`
+
+[[ -e $DIR/host.id ]] && { echo "Container already running."; exit 1; }
 
 trap 'rm -f $DIR/host.id' EXIT INT TERM HUP
 
@@ -23,4 +25,4 @@ TTY=`test "$INPUT_FILE" = "/dev/stdin" && echo "-t" || echo ''`
 # RUN_CMD="/files/start.sh"
 
 docker run $TTY -a stdout -a stdin --cidfile=$DIR/host.id $VOLUMES -P -i --rm -w "/files" \
-  $DENV --hostname $CURRENT_DIR $DOCKERINDEX$CURRENT_DIR ${RUN_CMD:""} < "$INPUT_FILE" > "$OUTPUT_FILE"
+  $DENV --hostname $CONTAINER $DOCKERINDEX$CONTAINER ${RUN_CMD:""} < "$INPUT_FILE" > "$OUTPUT_FILE"
